🎨 Refactor bin_werewolf
